Fix spacing and font size overmeasure
It was too late when I found out that the whole time before this point,
the live display of the page was zoomed in, and so things broke when I
zoomed out to default. But I'm keeping the font size larger than in the
provided picture demo in the project instructions as this way is more
accessible and readable for people that look at large screens with faces
a meter away from the screen.